+++ packages/azure-functions-durable/src/worker.ts
@@ -4,8 +4,8 @@
 import { TaskHubGrpcWorker, TaskHubGrpcWorkerOptions } from "@microsoft/durabletask-js";
 
 export class DurableFunctionsWorker extends TaskHubGrpcWorker {
-  constructor(options: TaskHubGrpcWorkerOptions = {}) {
-    super(options);
+  constructor(options: Omit<TaskHubGrpcWorkerOptions, "maximumTimerIntervalMs"> = {}) {
+    super({ ...options, maximumTimerIntervalMs: undefined });
   }
 
   async handleOrchestratorRequest(encodedRequest: string): Promise<string> {